move more stuff and add tests
